refactor: separate the surface chemistry classes
refactor: separate the surface chemistry related codes in Chemistry class to a separate surface_chemistry class. move the Material and the SurfacePhase classes to a separate module surface_components.
